ref(profile): split components to match file architecture
