+++ components/Profile/EmailVerification/Section.tsx
@@ -1,0 +1,3 @@
+export const EmailVerificationSection = () => {
+    return <></>
+}
